Implemented transactional pattern matching. #361
Graphs are transformed into transactions and their elements matched for a specific pattern. The results are provided for custom matching algorithms.

The default implemented matching algorithm uses depth-first search while growing a stack of embeddings.

Via a boolean flag, it is possible to specify if the results shall be the input graphs with a new property that determines if this graph contains the pattern or not, or the results are a collection of all found embeddings with a lineage property.