Concerning the Equality Test, there were some errors concerning the one provided within the project.
At first, two strings that appears to be the different they actually
describe the same graph. So, since the equality is string based, it
failed. So I had to change the test in order to check the actual vertices
and edges that really had matched.

By the way, by doing so all the Fuse operator for issue #491 is OK.
* All test are passed (with the new testing function)
* All tests are coded (I implemented even Martin's two examples)
* The fuse operator is defined completely