Probable typo in SNA example yields empty result
In the example SNA files provided in gradoop/gradoop-examples/src/main/resources/data/json/sna the label "Person" starts with a capital P. Using a small "p" in the getLabel().equals() predicate makes the example compile and run correctly, but the final result is empty. I suppose this was not the intended behaviour.